refactor(graph): move graph input helpers
